rename components to ui, fix loading feed
