+++ src/ui/atoms/image.tsx
